feat: Extract multi-line chart component for reuse
